# [MRG] Memory and speed improvements for colourspace conversion