#### Describe the changes
* Switch from dot product to matrix multiplication so numpy can use BLAS
* Ensure 32-bit floats are used
* Add option to allow frame-by-frame processing
* Try and reuse array memory where possible
* Add benchmarks

Related to #1348

Before:
```
TimeConversion.time_rgb_ybr_32_3_2f            157±0ms
TimeConversion.time_ybr_rgb_32_3_2f            212±0ms
```

After:
```
TimeConversion.time_rgb_ybr_32_3_2f            129±0ms
TimeConversion.time_ybr_rgb_32_3_2f            127±0ms
```

#### Tasks
- [x] Benchmarks added that prove feature is working
- [x] Fix or feature added
- [x] Unit tests passing and overall coverage the same or better