@pep8speaks 2021-04-13
Hello @scaramallion! Thanks for updating this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:

* In the file [`pydicom/benchmarks/bench_handler_colourspace.py`](https://github.com/pydicom/pydicom/blob/f6c1f377cade9c6a1408686027fe52ff0f672992/pydicom/benchmarks/bench_handler_colourspace.py):

> [Line 39:80](https://github.com/pydicom/pydicom/blob/f6c1f377cade9c6a1408686027fe52ff0f672992/pydicom/benchmarks/bench_handler_colourspace.py#L39): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)
> [Line 44:80](https://github.com/pydicom/pydicom/blob/f6c1f377cade9c6a1408686027fe52ff0f672992/pydicom/benchmarks/bench_handler_colourspace.py#L44): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)